move label_studio specifics to dedicated script
